backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
Unipath==1.1
